+++ tests/test_GridView.cpp
@@ -11,7 +11,7 @@
 
 TEST_CASE("test get_min_max_elevation")
 {
-    Chorasmia::Array2D<double> values({-999, -1, 4, 3, 1, -4}, 2, 3);
+    Chorasmia::Array2D<float> values({-999, -1, 4, 3, 1, -4}, 2, 3);
     GridLib::Grid grid(std::move(values));
     grid.set_unknown_elevation(-999);
     auto [min, max] = GridLib::get_min_max_elevation(grid.view());
@@ -21,7 +21,7 @@ TEST_CASE("test get_min_max_elevation")
 
 TEST_CASE("test get_bounding_rect")
 {
-    Chorasmia::Array2D<double> values({-999, -1, 4, 3, 1, -4}, 2, 3);
+    Chorasmia::Array2D<float> values({-999, -1, 4, 3, 1, -4}, 2, 3);
     GridLib::Grid grid(std::move(values));
     grid.set_planar_coords(GridLib::PlanarCoords{500000, 6000000});
     grid.set_row_axis({{10, 0, 0}, GridLib::Unit::METERS});
